==> seccao-1-pandas-hands-on/2-dataframe.ipynb <==
import pandas as pd

dados = {
    'Nome': ['Alice', 'Bob', 'Charlie'],
    'Idade': [24, 30, 22],
    'Cidade': ['São Paulo', 'Rio de Janeiro', 'Belo Horizonte']
}

df = pd.DataFrame(dados)

df

# Traz as 5 primeiras linhas se não passar valor
df.head(1)

# Traz as 5 últiomas linhas se não passar valor
df.tail(2)

dados_notas = {
    'Nome':['Alice', 'Bob', 'Charlie', 'David', 'Eva'],
    'Matemática':[85, 90, 78, 88, 92],
    'Fisíca':[80, 75, 88, 85, 95],
    'Química':[78, 82, 85, 80, 88]
}

df_notas = pd.DataFrame(dados_notas)

df_notas

# Calculando a média das notas por aluno
df_notas['Média'] = df_notas[['Matemática', 'Fisíca', 'Química']].mean(axis=1)
df_notas

# Indentificar alunos aprovados com nota boa (média >= 82.5)
alunos_aprovados = df_notas[df_notas['Média'] >= 82.5]
alunos_aprovados

# Exibir a média e os alunos aprovados
df_notas[['Nome', 'Média']]


alunos_aprovados[['Nome', 'Média']]
