==> seccao-1-pandas-hands-on/5-importar_csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv("../dados/produtos.csv", sep=',')
df_produtos

# Pega os 5 primeiros registros, se colocar o valor expecifico dentro do () ele também
df_produtos.head()

# Pega os 5 ultimos registros, se colocar o valor expecifico dentro do () ele também
df_produtos.tail(10)

# Mostra quantas linhas e colunas contém no DataFrame
df_produtos.shape

# Mostra os tipos de dados contidos no DataFrame
df_produtos.dtypes

df_produtos.sort_values(by='Quantidade', ascending=False)

df_produtos['Categoria'].value_counts()
