==> seccao-1-pandas-hands-on/6-analise_csv.ipynb <==
import pandas as pd

df_produtos = pd.read_csv("../dados/produtos.csv", sep=',')
df_produtos

df_produtos.dtypes

df_produtos['Quantidade'] = pd.to_numeric(df_produtos['Quantidade'])
df_produtos['Preço'] = pd.to_numeric(df_produtos['Preço'])

# 1-Desempenho de Vendas por Categorias
vendas_categoria = df_produtos.groupby('Categoria')['Quantidade'].sum()
vendas_categoria

# 2-Produto mais vendido
produto_mais_vendido_nome = df_produtos.groupby('Produto')['Quantidade'].sum().idxmax()
produto_mais_vendido_nome

# 3-Total de Unidades Vendidas por Produto
total_unidades_produtos = df_produtos.groupby('Produto')['Quantidade'].sum()
total_unidades_produtos

# 4-Preço Médio por Produto
media_preco_produto = df_produtos.groupby('Produto')['Preço'].mean()
media_preco_produto

# 5-Correlação entre Unidades Vendidas e Preço Unitário
correlacao = df_produtos['Quantidade'].corr(df_produtos['Preço'])
correlacao

vendas_categoria_df = vendas_categoria.reset_index()
vendas_categoria_df.columns = ['Categoria', 'UnidadesVendidas']

vendas_categoria_df

vendas_categoria_df.to_csv("../dados/vendas_categoria.csv", index=False)
