==> seccao-1-pandas-hands-on/8-analise_planilha.ipynb <==
import pandas as pd

df = pd.read_excel("../dados/campeonato_futebol.xlsx")
df

df.dtypes

df.describe()

df.head(4)

df.tail(4)

# 1-Análise de Desempenho por Time
desempenho_info = df[['Times', 'Pontos', 'Vitorias', 'Saldo_Gols', 'Gols_Pro', 'Gols_Contra']].set_index('Times')
desempenho_info

# 2-Time com mais vitória
time_mais_vitoria = df[df['Vitorias'] == df['Vitorias'].max()]
time_mais_vitoria

# 3-Time com menos vitoria
time_menos_vitoria = df[df['Vitorias'] == df['Vitorias'].min()]
time_menos_vitoria

# 4-Distribuição de GOLs Pró
df[['Times', 'Gols_Pro']].sort_values(by='Gols_Pro', ascending=False)


# 5-Relação entre gols pró e pontos
correlacao = df['Gols_Pro'].corr(df['Pontos'])
correlacao
