==> seccao-1-pandas-hands-on/9-cria_json.ipynb <==
import json

dados_receitas = {
    "Receitas": [
        {
            "titulo": "Bolo de Chocolate",
            "tempo_preparo": "45 minutos",
            "porcoes": 8,
            "ingredientes": [
                {"ingredientes": "Farinha de Trigo", 'quantidade': "2 xícaras"},
                {"ingredientes": "Açúcar", 'quantidade': "1 xícaras"},
                {"ingredientes": "Ovos", 'quantidade': "3 xícaras"},
                {"ingredientes": "Leite", 'quantidade': "1 xícaras"},
                {"ingredientes": "Manteiga", 'quantidade': "1/2 xícaras"}
            ],
        },
        {
            "titulo": "Salada Caesar",
            "tempo_preparo": "20 minutos",
            "porcoes": 4,
            "ingredientes": [
                {"ingredientes": "Alface romana", "quantidade": "1 cabeça"},
                {"ingredientes": "Croutons", "quantidade": "1 xícara"},
                {"ingredientes": "Queijo parmesão", "quantidade": "1/2 xícara"},
                {"ingredientes": "Molho caesar", "quantidade": "1/4 xícara"},
            ],
        },
        {
            "titulo": "Pasta ao Pesto",
            "tempo_preparo": "30 minutos",
            "porcoes": 2,
            "ingredientes": [
                {"ingredientes": "Massa", "quantidade": "200g"},
                {"ingredientes": "Manjericão fresco", "quantidade": "1/2 xícara"},
                {"ingredientes": "Azeite de olíva", "quantidade": "1/3 xícara"},
                {"ingredientes": "Nozes", "quantidade": "1/4 xícara"},
                {"ingredientes": "Queijo parmesão", "quantidade": "1/4 xícara"},
            ],
        }
    ]
}

with open("../dados/receitas.json", "w", encoding="utf-8") as file:
    json.dump(dados_receitas, file, ensure_ascii=False, indent=4)
