==> seccao-1-pandas-hands-on/11-analise_json.ipynb <==
import pandas as pd

df = pd.read_json("../dados/receitas.json", encoding="utf-8")
df

# Análise Total de Ingredientes por Receita
receitas_expandidas = []

for receita in df['Receitas']:
    total_ingredientes = len(receita['ingredientes'])
    receita_info = {
        'titulo': receita['titulo'],
        'tempo_preparo': receita['tempo_preparo'],
        'porcoes': receita['porcoes'],
        'total_ingredientes': total_ingredientes
    }
    receitas_expandidas.append(receita_info)

receitas_expandidas

df_expandido = pd.DataFrame(receitas_expandidas)
df_expandido
